ptp_clock: fix build on windows
c_ulong/gulong can be 32bit or 64bit depending on the platform.
Always cast it to u64 so the compiler will be happy.

+++ gstreamer-net/src/ptp_clock.rs
@@ -88,7 +88,7 @@ impl PtpClock {
             );
             assert_ne!(id, 0);
 
-            PtpStatisticsCallback(NonZeroU64::new_unchecked(id))
+            PtpStatisticsCallback(NonZeroU64::new_unchecked(id as u64))
         }
     }
 }
